requirements-dev.txt
-r requirements.txt
mypy
django-stubs
black
pytest-django
